🧬 Migrate the *template* from module level as sub-package level.
